refacto post Adrien Morin
