+++ AllWeShare/.env
@@ -22,5 +22,5 @@ DATABASE_URL=pgsql://user:password@127.0.0.1:5432/allweshare
 # For Gmail as a transport, use: "gmail://username:password@localhost"
 # For a generic SMTP server, use: "smtp://localhost:25?encryption=&auth_mode="
 # Delivery is disabled by default via "null://localhost"
-MAILER_URL=null://localhost
+MAILER_URL=gmail://allwesharethebest@gmail.com:AllWeShare2019@localhost?encryption=tls&auth_mode=oauth
 ###< symfony/swiftmailer-bundle ###
