+++ AllWeShare/src/Controller/Front/PostController.php
@@ -37,7 +37,7 @@ class PostController extends AbstractController
         return $this->render('Front/post/index.html.twig', [
             'post' => $post,
             'form' => $form->createView(),
-            'posts' => $postRepository->findAll(),
+            'posts' => $postRepository->findBy([], ['createdAt' => 'DESC']),
         ]);
     }
