+++ AllWeShare/src/Controller/Front/SecurityController.php
@@ -81,7 +81,7 @@ class SecurityController extends Controller
 
             $options_mail = array(
                 'name' =>  $user->getFirstname(),
-                'token' => $_SERVER['HTTP_HOST'].'/activate/'.$user->getToken()
+                'token' => 'http://'.$_SERVER['HTTP_HOST'].'/activate/'.$user->getToken()
             );
 
             $mailSending->sendEmail('Hello New Sharer',
@@ -162,7 +162,7 @@ class SecurityController extends Controller
 
                 $options_mail = array(
                     'name' =>  $user_exist->getFirstname(),
-                    'token' => $_SERVER['HTTP_HOST'] .'/change_password/'.$user_exist->getToken()
+                    'token' => 'http://'.$_SERVER['HTTP_HOST'] .'/change_password/'.$user_exist->getToken()
                 );
                 $mailSending->sendEmail('Password Change',
                     $user->getEmail(),
