+++ AllWeShare/templates/Front/user/account.html.twig
@@ -28,7 +28,7 @@
                                 {{ form_widget(form.firstname) }}
                             </div>
                         </td>
-                        <td><input class="editValues btn button secondary" type="button" value="Edit"></td>
+                        <td><input class="editValues btn btn-outline-primary" type="button" value="Edit"></td>
                     </tr>
                     <tr  style="background-color: rgba(0, 0, 0, 0.05);">
                         <th>Lastname</th>
@@ -38,7 +38,7 @@
                                 {{ form_widget(form.lastname) }}
                             </div>
                         </td>
-                        <td><input class="editValues btn button secondary" type="button" value="Edit"></td>
+                        <td><input class="editValues btn btn-outline-primary" type="button" value="Edit"></td>
                     </tr>
                     <tr>
                         <th>Email</th>
@@ -52,7 +52,7 @@
                                 {{ form_widget(form.address) }}
                             </div>
                         </td>
-                        <td><input class="editValues btn button secondary" type="button" value="Edit"></td>
+                        <td><input class="editValues btn btn-outline-primary" type="button" value="Edit"></td>
                     </tr>
                     <tr  style="background-color: rgba(0, 0, 0, 0.05);">
                         <th>City</th>
@@ -62,12 +62,12 @@
                                 {{ form_widget(form.city) }}
                             </div>
                         </td>
-                        <td><input class="editValues btn button secondary" type="button" value="Edit"></td>
+                        <td><input class="editValues btn btn-outline-primary" type="button" value="Edit"></td>
                     </tr>
                     </thead>
                 </table>
 
-                <button class="btn button primary float-right">{{ button_label|default('Save') }}</button>
+                <button class="btn btn-primary float-right">{{ button_label|default('Save') }}</button>
 
             {{ form_end(form) }}
             &nbsp;
