+++ AllWeShare/templates/base.html.twig
@@ -33,7 +33,7 @@
 
     {{ encore_entry_script_tags('app') }}
 
-    <script src="{{ asset('js/bootstrap.js"') }}"></script>
+    <script src="{{ asset('js/index.js') }}"></script>
 {% endblock %}
 </body>
 </html>
