+++ AllWeShare/templates/navigation/navbar.html.twig
@@ -1,7 +1,7 @@
 <nav class="navbar navbar-expand-lg bg-light">
     <div class="col-md-2">
         <a class="navbar-brand" href="{{ path('post_index') }}">
-            <img src="{{ asset('images/allweshare-logo.png') }}" class="img-fluid"width="166" height="166">
+            <img src="{{ asset('images/allweshare-logo.png') }}" class="img-fluid"width="100" height="100">
         </a>
     </div>
     <div class="collapse navbar-collapse">
